changed ValueData from byte to float

+++ Ported/BucketBrigade/Assets/Scripts/Components/ValueComponent.cs
@@ -1,9 +1,10 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using Unity.Entities;
+using Unity.Mathematics;
 using UnityEngine;
 [GenerateAuthoringComponent]
 public struct ValueComponent : IComponentData
 {
-    public byte Value;
+    public float Value;
 }
